:art:新增user app,调整原有代码结构
